requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.7
passlib==1.7.4
bcrypt==3.2.2
PyJWT==2.8.0
python-docx==1.1.0
openpyxl==3.1.2
pdfplumber==0.10.3
pygetwindow==0.0.9
pywinauto==0.6.9
pywin32==307
openai==1.30.1
python-dotenv==1.0.1 